Rework MouseLightAPI
Extract the filtering part from the fetch_neurons_metadata() method so it
is possible to filter a list of metadata without calling the API again.
Fix the filter to keep the neurons whose soma is inside one of the given
region instead of only in their ancestors.
Add a function to fetch an atlas with a retrying feature because the atlas
server is quite unstable.
Fix cache path to actually use the MouseLightAPI cache instead of the
NeuroMorpOrgAPI cache.